nbody: Remove old, now inaccurate comment

+++ nbody/include/nbody.h
@@ -21,14 +21,16 @@ along with Milkyway@Home.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef _NBODY_H_
 #define _NBODY_H_
 
+#include "nbody_config.h"
+#include "nbody_types.h"
+#include "milkyway_util.h"
+
+#include <time.h>
+
+
 #ifdef _cplusplus
 extern "C" {
 #endif
-
-#include "nbody_config.h"
-#include "nbody_types.h"
-
-#include <time.h>
 
 /* Command line arguments */
 typedef struct
@@ -42,7 +44,7 @@ typedef struct
     real* serverArgs;
     unsigned int numServerArgs;
 
-    long setSeed;         /* the PRNG uses a long for a seed, but int is more portable. */
+    uint32_t setSeed;
     int outputCartesian;
     int printTiming;
     int verifyOnly;
